Special considerations for constants defined by Tactician itself
We have a very annoying chicken-egg problem here. Tactician defines a constant
for private use. However, we cannot (easily) generate a .bin file for these
constants because this plugin has not been compiled yet at that point. In normal
circumstances, this constant is not really used anyway, so we just filter it out
from the global context. However, for the rare circumstances that something
refers to this constant anyway, we allow the constant to be inlined into other
.bin files.